+++ tests/py/query/manager.py
@@ -31,16 +31,16 @@ class TestManager(test.TestCase):
     
     def testGetOrCreate(self):
         objects = self.mapper[SimpleModel]
-        v,  created = yield objects.get_or_create(code='test')
+        v,  created = yield objects.update_or_create(code='test')
         self.assertTrue(created)
         self.assertEqual(v.code,'test')
-        v2, created = yield objects.get_or_create(code='test')
+        v2, created = yield objects.update_or_create(code='test')
         self.assertFalse(created)
         self.assertEqual(v,v2)
         
     def test_get(self):
         objects = self.mapper[SimpleModel]
-        v, created = yield objects.get_or_create(code='test2')
+        v, created = yield objects.update_or_create(code='test2')
         self.assertTrue(created)
         v1 = yield objects.get(code='test2')
         self.assertEqual(v1, v)
@@ -62,21 +62,21 @@ class TestManager(test.TestCase):
     def testUniqueFilter(self):
         objects = self.mapper[SimpleModel]
         yield self.async.assertEqual(objects.filter(code='test4').count(), 0)
-        yield objects.get_or_create(code='test4')
+        yield objects.update_or_create(code='test4')
         yield self.async.assertEqual(objects.filter(code='test4').count(), 1)
         yield self.async.assertEqual(objects.filter(code='foo').count(), 0)
         
     def testIndexFilter(self):
         objects = self.mapper.simplemodel
         yield self.async.assertEqual(objects.filter(group='g1').count(), 0)
-        v, created = yield objects.get_or_create(code='test5', group='g2')
+        v, created = yield objects.update_or_create(code='test5', group='g2')
         yield self.async.assertEqual(objects.filter(group='g1').count(), 0)
         yield self.async.assertEqual(objects.filter(group='g2').count(), 1)
         v1 = yield objects.get(group='g2')
         self.assertEqual(v, v1)
         yield self.async.assertRaises(SimpleModel.DoesNotExist,
                                       objects.get, group='g1')
-        v2, created = yield objects.get_or_create(code='test6', group='g2')
+        v2, created = yield objects.update_or_create(code='test6', group='g2')
         yield self.async.assertEqual(objects.filter(group='g2').count(), 2)
         yield self.async.assertRaises(stdnet.QuerySetError,
                                       objects.get, group='g2')